+++ pybursa/urls.py
@@ -2,10 +2,11 @@ from django.conf.urls import patterns, include, url
 from django.contrib import admin
 
 from pybursa import views
+from feedbacks.views import FeedbackView
 
 urlpatterns = patterns('',
     url(r'^$', views.index, name='index'),
-    url(r'^feedback/', views.FeedbackView.as_view(), name='feedback'),
+    url(r'^feedback/', FeedbackView.as_view(), name = "feedback"),
     url(r'^quadratic/', include('quadratic.urls')),
     url(r'^contact/', views.contact, name='contact'),
     url(r'^student_list/', views.student_list, name='student_list'),
