+++ students/views.py
@@ -13,6 +13,7 @@ from django.views.generic.detail import DetailView
 
 class StudentListView(ListView):
     model = Student
+    paginate_by = 2
 
     def get_queryset(self):
         course_id = self.request.GET.get('course_id', None)
