+++ templates/students/student_list.html
@@ -66,5 +66,26 @@
               {% endif %}  
             </tbody>
            </table>
+            <nav>
+              <ul class="pagination">
+                <li>
+                  {% if page_obj.has_previous %}
+                  <a href="?page={{ page_obj.previous_page_number }}"><< previous</a>
+                  {% endif %} 
+                </li>
+                {% for page in paginator.page_range %}    
+                  {% if page == page_obj.number %}
+                    <li class="active"><a href="?page={{ page }}">{{ page }}</a></li>
+                  {% else %}
+                    <li><a href="?page={{ page }}">{{ page }}</a></li>
+                  {% endif %}
+                {% endfor %} 
+                <li>
+                  {% if page_obj.has_next %}
+                   <a href="?page={{ page_obj.next_page_number }}">next >></a>
+                  {% endif %}
+                </li>
+              </ul>
+            </nav>
       </div>        
 {% endblock content %}
